Fixed better-acquisition-yielding
Forgot to access timeRemaining as a function, to properly compare and hold off on acquiring an investment if a better one is approaching completion. Also returned check to 50% of considered investment target time, rather than 100%, to hopefully squeeze more investments in, since that didn't seem to be the real issue.