Added risk-capping to R&D
Default implementation defaults to 0.0% risk but is somewhat lazy at this edge-case; it doesn't check to see if hiring will push risk above allowable threshold, so there will be some minimal nonzero risk when it gets done hiring. This implementation does not ever fire employees in order to reduce risk.