Idle Class v0.6.0 Compatibility Update
Just a tweak to the index for the current bankruptcy multiplier. It's the only stat in the game where I haven't been able to track down a handy in-game function to grab it. So any changes to the game's stats array can potentially throw this off. Also refactored a bit.